+++ src/pages/MyNfts/MyNFTsPage.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import { Loader, ErrorMessage, BookCard, AppButton } from '@/common'
-import MyNftsNoData from '@/pages/my-nfts/MyNftsNoData.vue'
+import { MyNftsNoData } from '@/pages/MyNfts'
 
 import { ErrorHandler } from '@/helpers'
 import { ref, watch } from 'vue'