+++ src/pages/MyNfts/index.ts
@@ -1,0 +1,1 @@
+export { default as MyNftsNoData } from '@/pages/MyNfts/MyNftsNoData.vue'