+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Network switcher component in bookshelft(doing nothing for now)
 - Network switcher component in header
 - User drop-down menu
+- useContext hook instead of calling useI18n every time and specifying scope
 
 #### Changed
 - About page redesign
@@ -22,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Removed
 - Metamask disconnect feature
+- Unused login form
 
 #### Fixed
 - Too large book title now displays correct on all pages